Play endpoint returns headers required for range requests
The Accept-Ranges response HTTP header is a marker used by the server to advertise its support of partial requests. The value of this field indicates the unit that can be used to define a range.
The Content-Length entity header indicates the size of the entity-body, in bytes, sent to the recipient.
The Content-Range response HTTP header indicates where in a full body message a partial message belongs.