Don't GZIP audio streams, ever.
HTTP4K does not support zipping streams.
An inconsistent behaviour was observed where streaming the audio via an HTML audio element was OK, but a GET request to the same resource would hang and error.
It transpired that the web audio element set the Accept-Encoding header to:
Accept-Encoding: identity;q=1, *;q=0
whereas the normal browser behavious is to set:
Accept-Encoding: gzip, deflate, br

Rather than try to modify the header in the browser, I am preferring to ignore the header for the `play` request.

Unfortunately that means that I can't test the whole filter stack in quite the same way, but the dummy app in the filter tests has been organised in the same way as the real app. This revealed that the error pages would no longer be Gzipped as the Gzip filter is not applied to every single endpoint. Rather than write my own Gzip filter I will live with this trade-off - shouldn't be a big deal in the real world, and it's documented in a test.