Download ffmpeg and ffprobe during build if not available locally.
I want to get these out of version control, however there are few things that aren't that nice here.

Firstly, the files are large - approx 70 MB each. This means the build is significantly slower. If a Docker image could be cached on CI this might be a faster option (though if the image has to be downloaded every time it might be slower).

Secondly, there's lots of duplication in this script that could be factored out.